Bug 794303 - Declare m_totalMsgSize as uint64_t instead of uint32_t for dealing with larger than 4GB mail folders. r=irving

+++ mailnews/base/src/nsMsgFolderCompactor.h
@@ -62,7 +62,7 @@ protected:
   uint32_t m_size;
 
    // sum of the sizes of the messages, accumulated as we visit each msg.
-  uint32_t m_totalMsgSize;
+  uint64_t m_totalMsgSize;
 
   uint32_t m_curIndex; // index of the current copied message key in key array
   uint64_t m_startOfNewMsg; // offset in mailbox of new message
